+++ .github/pull_request_template.md
@@ -1,0 +1,40 @@
+# Description
+
+- Please include a summary of the change and which issue is fixed.
+- Please also include relevant motivation and context.
+- List any dependencies that are required for this change.
+
+
+- Fixes # (issue)
+## Type of change
+
+Check relevant points.
+
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] New feature (non-breaking change which adds functionality)
+- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] This change requires a documentation update
+
+# How Has This Been Tested?
+
+- Please describe the tests that you ran to verify your changes.
+- Provide instructions so we can reproduce.
+- Please also list any relevant details for your test configuration
+
+- [ ] Test A
+- [ ] Test B
+
+# Checklist:
+Prepare items below using:
+\[ :x: \] (markdown: `[ :x: ]`) for TODO items
+\[ :white_check_mark: \] (markdown: `[ :white_check_mark: ]`) for DONE items
+\[ N/A \] for items that are not applicable for this PR.
+
+
+- [ ] updated version number in setup.py/pyproject.toml/environment.yml.
+- [ ] updated the lock file.
+- [ ] added changes to History.rst.
+- [ ] updated the latest version in README file.
+- [ ] I have added tests that prove my fix is effective or that my feature works.
+- [ ] New and existing unit tests pass locally with my changes.
+- [ ] documentation are updated.
