+++ hydrolib/core/dflowfm/extold/models.py
@@ -164,6 +164,21 @@ class ExtOldBoundaryQuantity(StrEnum):
     """Discharge-water level dependency"""
 
 
+class ExtOldParametersQuantity(StrEnum):
+    """Enum class containing the valid values for the Spatial parameter category
+    of the external forcings.
+
+    for more details check D-Flow FM User Manual 1D2D, Chapter D.3.1, Table D.2
+    https://content.oss.deltares.nl/delft3d/D-Flow_FM_User_Manual_1D2D.pdf
+    """
+
+    FrictionCoefficient = "frictioncoefficient"
+    HorizontalEddyViscosityCoefficient = "horizontaleddyviscositycoefficient"
+    HorizontalEddyDiffusivityCoefficient = "horizontaleddydiffusivitycoefficient"
+    AdvectionType = "advectiontype"
+    InfiltrationCapacity = "infiltrationcapacity"
+
+
 class ExtOldMeteoQuantity(StrEnum):
 
     # Meteorological fields
